+++ Editor/TextureBlender.cs
@@ -2,22 +2,22 @@
 using UnityEngine;
 using System.Collections.Generic;
 using net.rs64.TexTransCore.BlendTexture;
+using System;
 namespace net.rs64.TexTransTool
 {
     [AddComponentMenu("TexTransTool/TTT TextureBlender")]
     public class TextureBlender : TextureTransformer
     {
-        public Renderer TargetRenderer;
-        public int MaterialSelect = 0;
+        public RelativeTextureSelector TargetTexture;
+
         public Texture2D BlendTexture;
         public Color Color = Color.white;
         public BlendType BlendType = BlendType.Normal;
-        public PropertyName TargetPropertyName;
 
 
-        public override List<Renderer> GetRenderers => new List<Renderer>() { TargetRenderer };
+        public override List<Renderer> GetRenderers => new List<Renderer>() { TargetTexture.TargetRenderer };
 
-        public override bool IsPossibleApply => TargetRenderer != null && BlendTexture != null;
+        public override bool IsPossibleApply => TargetTexture.TargetRenderer != null && BlendTexture != null;
 
         public override TexTransPhase PhaseDefine => TexTransPhase.BeforeUVModification;
 
@@ -25,19 +25,28 @@ namespace net.rs64.TexTransTool
         {
             if (!IsPossibleApply) return;
 
+            var DistTex = TargetTexture.GetTexture();
+            if (DistTex == null) { return; }
+
+            var AddTex = TextureBlendUtils.CreateMultipliedRenderTexture(BlendTexture, Color);
+            Domain.AddTextureStack(DistTex, new TextureBlendUtils.BlendTextures(AddTex, BlendType));
+        }
+    }
+    [Serializable]
+    public class RelativeTextureSelector
+    {
+        public Renderer TargetRenderer;
+        public int MaterialSelect = 0;
+        public PropertyName TargetPropertyName;
+
+        public Texture2D GetTexture()
+        {
             var DistMaterials = TargetRenderer.sharedMaterials;
 
-            if (DistMaterials.Length <= MaterialSelect) return;
+            if (DistMaterials.Length <= MaterialSelect) return null;
             var DistMat = DistMaterials[MaterialSelect];
 
-            var DistTex = DistMat.GetTexture(TargetPropertyName) as Texture2D;
-            var AddTex = TextureBlendUtils.CreateMultipliedRenderTexture(BlendTexture, Color);
-            if (DistTex == null)
-            {
-                return;
-            }
-
-            Domain.AddTextureStack(DistTex, new TextureBlendUtils.BlendTextures(AddTex, BlendType));
+            return DistMat.GetTexture(TargetPropertyName) as Texture2D;
         }
     }
 }