+++ MultLayerImage-Parser/LayerData/RasterLayerData.cs
@@ -1,7 +1,7 @@
 using System;
 using UnityEngine;
 
-namespace net.rs64.TexTransCore.Layer
+namespace net.rs64.MultiLayerImageParser.LayerData
 {
     [Serializable]
     public class RasterLayerData : AbstractLayerData